Player sprite, volume slider, camera changes
Added in player sprite. However it is only 2 frames of an 'idle' action, even when moving.

Volume slider works. Cannot fully test it since we don't have sound effects yet.

Made the camera its own object rather than being a child of the Player object, which was causing problems when the player would flip when moving right to left.