Save and load game mechanic
Added functional save and load buttons to the pause menu and changed the main menu to have a load game and a new game button. the save button saves the health, level, exp, and position of the player. load button replaces the players current data with that data and the new game button resets the whole scene when loading in.